Move GLibTypeDeleter.h to src/platform
