+++ src/platform/GLibTypeDeleter.h
